SplitParty_backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
email-validator==2.0.0.post3
python-multipart==0.0.6